requirements/base.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.79
botocore==1.27.79
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==2.2.5
Django==4.1.1
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-modeltranslation==0.18.4
django-phonenumber-field==7.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
phonenumbers==8.12.55
Pillow==9.2.0
pkg_resources==0.0.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
twilio==7.14.1
uritemplate==4.1.1
urllib3==1.26.12
